+++ src/dune_engine/subst_config.mli
@@ -1,0 +1,13 @@
+open! Stdune
+open Import
+
+type t =
+  | Disabled
+  | Enabled
+
+val to_dyn : t -> Dyn.t
+
+val field :
+  since:Dune_lang.Syntax.Version.t -> t option Dune_lang.Decoder.fields_parser
+
+val of_config : t option -> t